Reduced bundle size 91%; Added Chunking
