+++ client/TitleScreen/TitleScreen.html
@@ -1,4 +1,11 @@
 <div class="SplashScreen-container">
+	<img
+		src="https://github.com/F28WP-Dubai-Group-6/LunarLander/blob/master/Assets/Demo.gif?raw=true"
+		alt=""
+		class="background"
+		crossorigin="anonymous"
+	/>
+
 	<div class="gradient-filter"></div>
 	<div class="logo-container">
 		<div class="main-logo-text">
