+++ client/src/Assets/Sounds/Radio/Before the night/BeforeTheNight.js
@@ -1,3 +1,3 @@
 export async function getSong(index, name) {
-	return await import(/* webpackMode: "eager" */ `./${index}- ${name}.mp3`);
+	//return await import(/* webpackMode: "eager" */ `./${index}- ${name}.mp3`);
 }
